fix: hot loader with native
