+++ native.js
@@ -1,24 +1,24 @@
 import { AppRegistry } from 'react-native'
-import React from  'react'
+import React from 'react'
 //need to import this here before webviewcrypto to prevent circular require because of
 //global.crypto = {} dont know why
 import 'react-native-crypto'
 import WebviewCrypto from '@gooddollar/react-native-webview-crypto'
 import Config from 'react-native-config'
-import App from './src/App'
+import App from './src/mainApp/AppHolder'
 import { name as appName } from './app.json'
 import 'react-native-gesture-handler'
 import { SafeAreaProvider } from 'react-native-safe-area-context'
 import './src/pushNotifications'
 import './src/lib/utils/deepLinking'
 
-const DeApp  = () => (
+const DeApp = () => (
   <SafeAreaProvider>
-    <WebviewCrypto/>
+    <WebviewCrypto />
     <App />
   </SafeAreaProvider>
 )
 
-console.disableYellowBox = !!Config.TEST_REACT_NATIVE;
+console.disableYellowBox = !!Config.TEST_REACT_NATIVE
 
 AppRegistry.registerComponent(appName, () => DeApp)
