+++ src/mainApp/AppHot.native.js
@@ -1,0 +1,1 @@
+export { App as default } from './App'
