+++ src/mainApp/AppHot.web.js
@@ -1,0 +1,4 @@
+import { hot } from 'react-hot-loader/root'
+import { App } from './App'
+
+export default hot(App)
